[IMP] web_studio: Allow editing of default group by for ListView
This commit adds a "Default Group By" field in the list view editor
sidebar, allowing the user to edit the default group by.

The group by is not applied inside of the studio view so the user can
keep editing it. When a default goup by field is selected, a message
informs the user that the changes are applied even though they are not visible.

Task ID: 3259513

closes odoo/enterprise#39582

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>