[IMP] timesheet_grid: make task/project field no_create_edit and not no_open
Before this commit project and task field is was possible to create and edit
task/project it was not good cause user create any type of project and task
from there even non allowed timesheet project and not opening project and
task it not providing any info of name of record if lots of records
have similar name.

This commit make task/project no_create_edit and make not no_open.

task-3251672

Part-of: odoo/enterprise#39936