[IMP] web_studio: add currency choice in monetary field
This commit improve the monetary field by changing the way currencies
are handled:

- The user can now change the currency in the monetary field.
- Now creating a new monetary without currencies in the model also cause
the creation of a new currency in the model and the view. Before, the
user was asked to create the currency field first.
- Now creating a new monetary with a model containing currencies cause
the usage of one of the currencies found. Before, the currency were not
detected if it's name were not `x_currency_id` or `currency_id`. The
currency is also now added to the view if it is not present.

The currency choice is saved in the monetary field in the backend side.
Meaning that we do not use the `currency_field` attribute in the arch:

- To have a better consistency for rounding: if `currency_field` is set
to a 3 decimal rounding in the arch and 2 decimal rounding in backend
then the backend has the final word as he save in the database by
rounding with 2 decimal.
- To have a better consistency between views: the same monetary in
different views will share the same currency.

closes odoo/enterprise#40597

Task-id: 3259535
Related: odoo/odoo#118199
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>