[FIX] website_sale_renting: add rental accessory product
The case where we want to add an accessory product
that can be rented on the ecommerce on the '/shop/cart' page
is not managed.

When we are on the '/shop/cart' page,
the "Add to cart" button will be displayed
if the product that triggered the proposal of this accessory
product is present in a line of the sale order.

Moreover, it is important to verify that
this line contains the time period.

If we can add the rentable accessory product to the cart,
we will apply the same time period as the product
that triggered this accessory product.

opw-3247134

closes odoo/enterprise#39451

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>