[FIX] knowledge: show AccessError when unfolding fails
Purpose:
--------
Currently, when unfolding an article in the sidebar, a "404 not found"
message is added under the row of the unfolded article if this article
is not accessible by the user anymore or if it has been deleted.

Now, an AccessError will be shown to the user, so that he knows why he
cannot unfold the article, and the article will be removed from the
sidebar.

Task-3266638

closes odoo/enterprise#39454

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>