[FIX] account_disallowed_expenses{_fleet}: fix group by vehicle
When you have multiple accounts in a category of Disallowed Expenses,
and you group by vehicle, the group by "hides" some of the accounts.
This should not happen of course.

The problem was caused by the key used during the postprocessing
of the query results. If the fleet rate was the same for several
accounts, the grouping key would be identical.
This meant that on the last value was shown since the previous
ones got overwritten.

This is now fixed and the tests have been adadpted accordingly
in order to cover this occurrence.

Also we should display percentages with a precision of two decimals.

task-3258972

closes odoo/enterprise#39465

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>