[IMP] website_helpdesk_*: make the help center beautiful
* = forum, knowledge, slides

This commit aims to make the help center better looking
by using separated lists of quick-links to items such as
forum posts, knowledge articles and courses.

Each list has at most 5 elements and they are sorted based on
votes / activity date for the forum posts, favourites / write
date for the knowledge articles and number of views for the
courses.

task-3444600

closes odoo/enterprise#45038

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>