[MOV] (website_)knowledge: move portal search tour to website_knowledge
This commit is a preparation commit for the upcoming feature that will allow
portal users to access the knowledge.articles using the OWL backend form view.

That means that portal users will no longer be redirected to the portal
readonly view, making the "knowledge_portal_search_tour" irrelevant for them.
It is however still relevant for *public* users when articles are published.

We thus move the code and the logic to the website_knowledge module and
slightly modify the test to publish the articles instead of granting access to
the portal user.

We also remove the part that was testing favoriting an article as public users
do not have access to that feature.

The core of the tour code, meant to test the search feature, is unchanged.

Task-3056280

Part-of: odoo/enterprise#38575