[FIX] industry_fsm: send multiple reports to a customer
Issue:
It is not possible to send multiple reports in Field Service
to a customer.
Only the first report will be sent.

Cause:
In the `_process_state` method in the flow of sending mail:
```py
...
mailing_document_based = self.env.context.get('mailing_document_based')
...
elif done_emails and mail_to in done_emails and not mailing_document_based:
    mail_values['state'] = 'cancel'
    mail_values['failure_type'] = 'mail_dup'
...
elif done_emails is not None and not mailing_document_based:
    done_emails.append(mail_to)
...
```
If the emails sent are not specified as document based,
there will be only one email per recipient.

Solution:
Specify in the context that there will be documents in the emails sent.

opw-3241212

closes odoo/enterprise#39889

X-original-commit: 4eca95bb11fdd6468d93ad786afb694645a34e94
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>