[FIX] knowledge: check wysiwyg before removing...
behaviors

Prior to this fix, we do not check that the wysiwyg was added before
trying to remove behaviors.

This leads to errors on unmounting if the field never started the
wysiwyg in the first place.
Like in cases where it is in xml or readonly mode.

Indirectly tested in: odoo/odoo#105406

task-3035163

closes odoo/enterprise#39537

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>