[FIX] account_sepa: display proper maximum amount when raising because of a too big amount
max_digits is used to count the number of digits in the string representation of the number, ignoring the coma. When displaying the error message, we remove 3 to this number to create the digits before the coma : 1 character for the coma, then 2 characters for the decimal places. It's wrong, and should be 2, as, again, the coma does not count in this limit.

closes odoo/enterprise#39603

X-original-commit: 7cd3ae29e8b735a3e0f5b37eaec0655a238fb121
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>