[FIX] account_intrastat: fix product weight/volume groups
A previous commit (c090562) added the "group_account_invoice" group to
the inventory page of the product view. This meant that, when
account_intrastat is installed, users with the account_invoice group
would be able to see and edit the weight and volume of a product (on the
inventory tab of the product form view).

However there is a typo (`seperator` should be `separator`), though this
doesn't actually matter since the default value of the separator in
odoo/tools/template_inheritance.py is a comma.

This commit corrects this typo (it is better to have it explicitly
defined) and an associated commit in stock will prevent this xpath from
being replaced (and instead add to the xpath).

closes odoo/enterprise#40101

X-original-commit: 2acb284c7df7b850e286d4363451e48f3dfda725
Related: odoo/odoo#119284
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>