[FIX] hr_contract_salary: Force assets generation
Purpose
=======

Currently the salary configurator won't load properly if called
before the frontend assets are generated.

As the template is rendered and the transaction is rollbacked before
loading the response to drop all the generated records used for the
simulation (new employee, contract, payslip, company car, etc...),
if the assets are generated with the template, they will be dropped
and used afterward, which is leading to a broken page.

For big production base like odoo.com this is not really an issue as
the frontend assets are always loaded, but this could be problematic
for smaller customer databases.

Specification
=============

The idea of this commit is to force the assets generation before
doing the simulation so that this use case is not occuring anymore.

closes odoo/enterprise#39562

Taskid: 3266112
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>