[IMP] sale_subscription: event_date should be the date of the log
Before this commit, the event_date of sale.order.log was either today or the
start_date of the order if it was not started yet. It had the advantage to avoid
displaying revenue for non started contracts (not invoiced yet) but it had drawbacks.

Some log would be created in the future for several usecases:
simple subscription starting in the future, upselled before started
transfer log on parent and children would have different event-date.
This would complicate the analysis as amount sum would be greater than 0.

Taskid: 3247318
Part-of: odoo/enterprise#40461