[FIX] pos_preparation_display: show unlinked preparation orders
It's possible to open a "Bar" without setting up the floors and tables. In this
situation, we still want to see that is being sent to the preparations display.

This commit fixes that behavior. First, we fix the traceback caused by sending
undefined 'pos_order_id' in creating the preparation display order records.
Then, when fetching the display orders, we also fetch those that are not linked
to pos.order.

closes odoo/enterprise#39729

X-original-commit: a704383429e0cfcdb35c963c1241172baa4f6672
Signed-off-by: Heinz Robin (rhe) <rhe@odoo.com>
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>