[FIX] planning_holidays: restrict the shifts fetched for _search_is_absent
Before this commit, when `is_absent` field is used in a domain to fetch
the `planning.slot` for which the employee took a day off during the
period of the slot, the `_search_is_absent` is called to return a domain
to fetch those slots. The problem with that method is all slots with
employee set are fetched to find the min and max datetime to search
time offs of all employees fetched with the slots. By doing that, if
many slots existed in the db, that search method will be really slow.

This commit fixes the issue by fetching only the slots in the future or
the ones containing the current date in their period. In the case, we
reduce the number of slots fetched in that to return the right domain.
The main dawnback of that changes is the search method will no longer
take into account the slots in the past. This problem is acceptable
since the use case of this field is to alert the planning manager,
an employee will not can do a slot because he took a day off and he
will have to reschedule the slot or assign it to someone else.

task-3269773

closes odoo/enterprise#39789

X-original-commit: 27487575d2bdbbbc799d9e66e1c6ce1a6eef6c86
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>