[IMP] social: stop unnecessary rendering on new post preview
Before this commit:
When creating a new post from social post form view and selecting some social
media platform then our post content will be rendered to all social platforms
(selected and not selected both) and a preview will be created for all, also
whenever we try to change content it will again render to all platforms.

After this commit, the content will be rendered only for the social media
platforms that we selected, so the preview loading time will decrease.

Task-3270260

closes odoo/enterprise#41199

X-original-commit: d7ea9cf96ddb6b5f7fdae9abf6eb0973e773617d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>