[FIX] sale_planning: leftover data on uninstall
On uninstall of sale_planning, planning slots linked to an SO line
would just have their SO line set to null, resulting in a planning
slot with neither start/end datetimes nor an SO line.

Upon reinstalling sale_planning, setting up the
`check_datetimes_set_or_plannable_slot` constraint would then fail, as
the previously created slots would be in an invalid state.

Cascade the deletion of SO lines into the planning slots, instead of
just setting the FK to NULL, this works for sale_planning demo data
SO, however it doesn't work when adding planning slots linked to
unrelated sale orders.

For that case, add an uninstall hook which removes the planning slots
which are *only* linked to an SO, and restore the constraints on the
old fields (which otherwise don't get restored during uninstall, only
at the next update of whichever module adds the constraints).

closes odoo/enterprise#39682

X-original-commit: 0588d4deeef4e9a6d30da4daf83848dcce56f5b0
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>