[FIX] l10n_cl_edi: add hard batching to fetch mail
Current behaviour:
The implementation of the `fetch_mail` function interogates
the email server used for the EDI based on the IMAP uids. When first
setting up the incoming email server record, the `l10n_cl_last_uid` used
to track the last maximum uid that was processesed is set to 1.
In pratice this leads to the function searching the IMAP server for mails
from 1:*, i.e. a list of ALL IMAP uids is returned and a `for` loop is used
to fetch one email at a time and validate the email attachements for invoice
data. `l10n_cl_last_uid` gets updated outside the for loop to the last
processed uid in the current run.
The problem arises when the client connects to a new email server already
containing thousand of emails. In that case, while trying to process all
emails, the process will time out on the standard cpu limit (900s). Thus
the `for` loop never finishes and `l10n_cl_last_uid` never updates. At each
new run, the function begins from uid=1 again and timing out again.

After this fix:
We add a conservative default batch size (hardcoded to 50), so that the
`for` loop will only loop over the next 50 queued emails. Given an average
processing time of 2-3 seconds for each fetched email, this should guarantee
that the process never times out on current hardware and default settings.
As the default recurrence for the fetchmail cron job is "every 5 minutes",
we can expect that in most real life scenarios the clients DB should sync
in a matter of days maximum.
We also add a `server.write({'l10n_cl_last_uid': new_max_uid})` to the
try block relating to the email processing. This will keep track more often
of the last max processed IMAP uid in case the job is aborted for some other
unforseen reason.

opw-3225345

closes odoo/enterprise#39685

X-original-commit: a8a9e2a5b0c644a442d5fd33c0f600353feb4d1f
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: jorv-odoo <jorv@odoo.com>