[FIX] knowledge: don't reload when refreshing article structure
Fixes 2 problems related to the articles structure command:

1) A Component should never modify its own props (only the parent can do it).
Using a state property for the content instead ensure that the component is
re-rendered when the value changes.

2) Calling a `record.save` may trigger a full reload of the view, it should be
done with `noReload: true` and `stayInEdition: true` to prevent that.

This commit is motivated by the fact that the test:
`Check Articles Index is correctly built - and updated` could fail (probably
because of the reload caused by the `record.save`).

Task-3236156

X-original-commit: a28b181dc34ecf39a1510f8cb448341894a18f43
Part-of: odoo/enterprise#39754