[FIX] l10n_de_pos_cert: prevent unexpected error in validating session
The _validate_session() method in the l10n_de_pos_cert module was
overridden without checking for a dictionary return value from the
super  method. This could cause unexpected behavior or errors when a
redirection or a wizard was returned.

For example, after calling the super function, it was possible that all
orders were in the "paid" state and the session validation was not
completed. This would cause the orders to be an empty recordset and
generate errors when generating a template.

After this commit it checks for a dictionary return value and returns
it immediately if one is found, preventing the rest of the code from
executing. This ensures that the code is only executed when it is
appropriate to do so.

opw-3265441

closes odoo/enterprise#39686

X-original-commit: 24a4c383853c8a5496031b1acaae06c0c533ffad
Signed-off-by: Heinz Robin (rhe) <rhe@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>