[FIX] currency_rate_live: misconfigured provider stops rate retrieval
When a provider does not support a company currency, a UserError is raised.
This is a problem for the cron, as all companies rate updates are stopped or rolled back.
To avoid this, we differentiate between the manual and cron using a context key.
In the case of manual update, errors are raised. But the cron will suppress UserError warnings to the log.

Task-3265345
OPW-3188333

closes odoo/enterprise#40336

X-original-commit: 187deb9779c1a5ccfc32db54c8ca59ffbda7f5ba
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>