[IMP] l10n_{*}_reports: growth comparison color
Improves the profit and loss reports of most of the localizations by properly displaying the color of comparison between periods. This enables a better understanding of comparison. The modified behavior aligns the generic P&L where expenses and depreciation consider a growth as negative. More precisely, we set the attribute green_on_positive to 'account report expressions' which displayed an expense or depreciation with a positive sign. Indeed, some P&L display these lines with a negative, and these did not require modification as the color was already the expected one.

closes odoo/enterprise#39977

Task: 3271475
Signed-off-by: John Laterre (jol) <jol@odoo.com>