[FIX] Use float_compare when selling assets.
When disposing/selling an asset, there is a special case in the
_compute_depreciation_value method of account_move that checks if we are
confronted with the disposal/sale move. This special case is entered based on a
comparison between an account_move_line 'balance' field, and
the 'original value' field of the asset.
Both of those fields are floats but are compared using a simple equality instead
of float_compare. In some rare occasion, that could result in a false negative
for that check.

Using float_compare instead of '=' fixes that issue.

This issue was originally detected in Odoo 16 but was first fixed in Odoo 14
and forward-ported to 16.

closes odoo/enterprise#40005

X-original-commit: 1958fa7b1bebe7590dd79c7d156dcc4d5eb481d7
Signed-off-by: William André (wan) <wan@odoo.com>