[FIX] helpdesk: fix traceback when helpdesk name is changed issue
Description of the issue/feature this PR addresses:
In helpdesk module, when name of helpdesk is try to update it gives traceback of
alias_name not found.

Current behavior before PR:
generate traceback when helpdesk name is updated.

Desired behavior after PR is merged:
helpdesk name is updated without traceback.

Fix:
alias_name is not found as there is no record with name alias_name in values
pass by _alias_get_creation_values function due to which traceback is generated.
we added condition as if alias_name is not in self then only its called
_alias_get_creation_values function.

task-3252521

closes odoo/enterprise#39757

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>