[FIX] data_merge, *: reinstall of modules with a data_merge_*
`action_merge_contextual_enable` has issues. Primarily the addition of
xids in odoo/enterprise#22664: because the xids are linked to
`data_merge` and this utility function is used by submodules, so when
e.g. `data_merge_crm` is uninstalled, the actions it has installed
remain, which is an issue.

However there is a more insidious issue: when the linked module (`crm`
above) is uninstalled, the model it defines are removed, which removes
the *server actions* via a cascade (on `model_id`).

However because the ir.model.data records are linked to `data_merge`,
they still are not removed when the *action* is deleted.

This, then, causes a major issue on reinstall: the model gets
reinstalled, the action has to follow, but when trying to re-create
the IMD that blows up because the same (module, name) already exists.

The hacky solution is to look for the existing IMD, and if it's found
update it to link to the action we just created.

The better solution would be to either create a record directly so it
can be linked to a XID hard-coded in the data file, or for the caller
to pass in their module name, so the action can be linked to the
correct module and garbage-collected when the module is uninstalled.

Sadly that would require data file changes at least and `-u`pdates for
it to work correctly, as well as an upgrade script to migrate the
existing xids over.

closes odoo/enterprise#39764

X-original-commit: 5ba1ce24ae860f71ab459fd5a83f77aed29d1901
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>