[IMP] account_reports: allow to add expressions to aged reports
The method _common_custom_unfold_all_batch_data_generator used to generate
a static dictionary according to the expressions loaded in `data/aged_partner_balance.xml`,
however, it was not possible to add new expressions due to the dictionary
generation couldn’t be inherited.

Proposed solution:

Creating a new method that generates and returns a dictionary. This improvement
allows us to inherit it and add new keys according to new expressions.

closes odoo/enterprise#44663

X-original-commit: d67db729a942eea83a8c7195e8a21a568b54da3d
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>