[FIX] knowledge: Make favorite button works in list view
Before this commit, the favorite widget was used to toggle the favorite
state for the current user. But the inverse method seems not called when using
the widget.

Anyway, a user can add an article in favorite even if the user has no write
access to the article. This is why we need to use the action_toggle_favorite
action button instead of a favorite widget on the field.

After this commit, the favorite toggle buttons in the list view properly
save the change.

Also, make sure we don't use inverse_is_user_favorite anymore by returning a
'Not Implemented Error', waiting for his complete removal in master.

Task-3274933

closes odoo/enterprise#39986

X-original-commit: e68d7b81a6403665355991e80ae5ab3e03260335
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>