[IMP] project: improve ux
Before this commit:

- project.updates 'hours planned' stat button: the
'plan orders' button take into account all SOs.
- Given a task assigned to user A and B.
If we schedule the task to user A, user B is removed.

After these changes:

- project.updates 'hours planned' stat button: the 'plan orders' button
should only take into account the SOs that are linked
to the project.
- When scheduling the task using the magnifying glass icon
of the Gantt view, the task should remain assigned to users
A and B if it is scheduled for either one of them.
If the task is scheduled for user C, the task should be
assigned to user C and users A and B should be removed from
the assignees.

task-2929871

closes odoo/enterprise#29911

Related: odoo/odoo#97220
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>