[FIX] account_reports: journal filter: multicompany issues with filter reload
To reproduce the issue (on a clean db with only the demo data):

1) Create a company A and B, each with a CoA installed
2) Create a journal group for company A, excluding any of its journals
3) Open the trial balance (or any other report with filter_journals set to true) with only company A active => your journal group is selected by default
4) Switch company to go to B while staying on the report
5) Switch again to A => the journal group is not selected by default this time

=> This happens because of the way the previous options are loaded. Switching the company still gives the options dictionary used by the report for the previous company as previous_options. So, in our case, when loading A for the second time, it will receive B's options as parameter.

The fix here is to apply additional conditions on the content of the previous options. If the journal information they contain don't overlap the journals available for the currently selected companies, it means the previous options were generated for another company, and we ignore them. This way, we'll reload A's first journal group in our example.

In the meantime, this commit fixes the way the active companies were treated by this report filter: self.env.companies won't match what we're computing the report for in case account.report's filter_multi_company field is set to 'disabled'. Only the multi_company option key should be used for that.

closes odoo/enterprise#39782

Signed-off-by: John Laterre (jol) <jol@odoo.com>