[IMP] l10n_be_hr_(payroll)_(contract_salary) : Add the possibility to have customized representation fees
Every company is not delivering the same advantages, according to that, the limit of representation fees you can consider as "with serious standard" is not the same for everybody, and it can vary depending on the job position.
Representation fees will now be divided as the sum of 4 parameters :
Homeworking amout : 145,81 euro
Internet amout : 30 euro
Phone amout : 25 euro
Car management amount : 80 euro

The user will now have the possibility to disable or enable each one of those parameters.

task - 3095671

closes odoo/enterprise#39790

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>