[IMP] sign: improve control of sign item option records
This commit adds a tree view for sign item options, changes the model to allow
archiving records. In order to prevent archiving from changing previous
templates or sign requests, it was decided to always show archived options
except when choosing an option while editing a template. For this reason, it
was decided not to use the active field, as we would need to use
test_active=False in a lot of places. Also, it was decided to drop archived
options when copying a template to prevent inconsistencies.

task-3233643

closes odoo/enterprise#39787

Related: odoo/odoo#122157
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>