[IMP] knowledge: add a button to print articles (to PDF)
This commit simply adds a shortcut to the browser "print" feature that will
allow end users to export their article to PDF (or print it).

We are not using "wkhtmltopdf" as it does not support emojis.
(And the investigation on that issue showed that adding the support for that
would be very complicated).

Note that exporting with the browser is not as good as we do not control
everything, for example the browser will add the URL of the page to the printed
document.

It is however possible to customize the export to some extent, if the end user
knows the various options.

In addition, we also dynamically load "print assets" before printing.
Those assets are not simply included in backend assets as they modify some
parts of the generic layout for a nicer article rendering to PDF.

Task-2892055

closes odoo/enterprise#39796

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>