[REF] knowledge: add loading spinners in the sidebar
Purpose:
-------
Currently, the data used to display the articles in the sidebar are fetched
in the `onWillStart` hook. As a result, the rendering of the article form
view is delayed until all the data used to display the sidebar are fetched.
If there are a lot of articles to show, the rendering of the view can take
some time, slowing down the navigation in the app.

After this commit, the data will be fetched in the setup method instead,
and the sidebar will display a spinner until the data are fetched.

A spinner will now also be shown when unfolding an article for the first
time since the loading of the page (in that case, the data of the child
articles need to be fetched), or when the article has a large number of
children (in that case, the rendering can take some time). This is meant
to replace the "load more" button which was introduced to prevent the app
from crashing when unfolding an article with a lot of articles in the same
subtree.

Task-3074123

Part-of: odoo/enterprise#34788