[FIX] web_studio: nested x2many edition works
In studio, navigate through a first x2many, then through a second.
The breadcrumb should now display: Form > SubView Form > SubView Form

Before this commit, there was a crash as the subview's xpath was not correctly computed.

After this commit, it works as expected, one can add a field, or edit in general,
the nested x2many.

closes odoo/enterprise#39803

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>