[IMP] account_reports: manage (in)active report variants
Some localizations might have multiple BS and P&L versions that can be
used for a company. Showing them all in the variant list selector makes
it hard for the user to choose, and also might mean the user has to
manually select the right version on every session.

In order to make it easier, we use the `active` field on reports.

We modify the report variant selector to only show active reports and
add an "Enable more ..." option that brings us to the report list
view where we can enable more or other variants of the current report.

This way, localizations with different BS and P&L can have only one
variant active by default, making sure the user is not confused and
always has the right version opened the first time.

task-3086473

Part-of: odoo/enterprise#35771