[FIX] knowledge: hide header buttons and banner when no record loaded
When the user does not have access to any article or there is no article
in the database, the editor view of Knowledge will not be able to load
an article and will display a no-content helper to invite people to
create a new article. With the recent changes, there are two issues
with the interface:

1. A banner appears above the no-content helper saying: "This article
   is only displayed to its members.". It makes no sens to display this
   banner as the view didn't load any record/article.
2. When hovering the "New" button of the header, the interface will
   reveal the buttons to add an icon, a cover or a new property field.
   Some of those buttons will issue a RPC call on the server using the
   id of the active record. As the view did not load any record, the id
   will be undefined and the user will get an error when interacting
   with those buttons.

This commit will hide the banner and the buttons "Add Icon", "Add Cover",
"Add Property" when the view didn't load any record. The issue has been
introduced unintentionally during the refactoring of the top bar. The
view will then look like how it used to be in Odoo 16.1.

Reference: https://github.com/odoo/enterprise/pull/35261
task-3254882

closes odoo/enterprise#40069

X-original-commit: d6326c5ba2cf757ba95f8e08fbd79f6431e41984
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>