[IMP] account_avatax: adds coordinates
Before this PR, we supported using the address of the Customer, but in some very
specific case, the address of the customer is not precise enough to get an
appropriate tax computation.
If latitude and longitude information are set on the partner then avalara will
use only the coordinates for computing the taxes.

closes odoo/enterprise#40403

Task-id: 3248767
X-original-commit: 5402ccd56cfc000ad34fa9525a67e97ef4d4e7e3
Signed-off-by: William André (wan) <wan@odoo.com>