[IMP] *_extract: update tracking message as OdooBot
There was an inconsistency about the author of the update message posted
in the chatter when the OCR filled the document.

If the user manually clicked on the "Refresh" button, they would be set
as the author, whereas if the refresh was performed automatically (i.e.
by the cron or the webhook), "Public User" would be the author.

Now, OdooBot will always be the author.

Community PR: odoo/odoo#117073

closes odoo/enterprise#38986

Signed-off-by: Arthur Gossuin (goa) <goa@odoo.com>