[FIX] project_enterprise: fix the planned_date removal of the tasks
This commit's purpose is to allow the user to remove the planned date
end, or the planned date start of a task from the list view, or the form
view of the task. If the user removes either the planned date end, or
the planned date start, both dates will be removed. If the user tries to
add only a date end or a date start while keeping the other date empty
on a task with no planned date, these changes will be discarded.
The goal of this fix is for the tasks to always have either both dates
set, or both dates set to False.

task-3254564

closes odoo/enterprise#39874

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>