[FIX] sale_subscription: cancel constraint was not checked before the state was set
Before this commit, the cancel constraint was check before the state was set.
The cancel constrait says that a canceled SO can't be in a subscription_state in
progress or paused or renewed.

As super was called after the status check, the subscription were put
in the 6_churn state before being canceled. It would prevent the constraint
to be catched. We call super before to prevent in progress subscription to be
cancelled without being properly closed.

Taskid: 3247318
Part-of: odoo/enterprise#40461