[IMP] sign: auto ordering roles
In sign it is possible to define an order for signature, however the current
process is really manual and there are some things that could be automated to
improve the usage of this feature. This commit adds a sequence number to the
sign.item.role model, which will be used for automatically defining the order
of signature. In case the order is not good, it's also possible to change it
manually.

task-3270748

closes odoo/enterprise#39890

Signed-off-by: Arnaud Joset <arj@odoo.com>