[IMP] knowledge: prevent double article in the breadcrumbs after invite
Purpose:
--------
Currently, after inviting someone to an article using the invite modal,
the article is added to the breadcrumbs even though we stay on the same
article (we thus have the same article name twice in the breadcrumbs).

This happens because the `action_invite_members` action returns an action
window to reload the form view (to reload the sidebar in case the article
went from private to shared).

After this commit, the action won't return an action window anymore, but
closing the invite modal will reload the permission panel and the sidebar
only if the article changed from category.

This also allows to keep the invite modal open after inviting new members
on an article, allowing the user to get a visual feedback that the members
have been invited.

Task-3159149

closes odoo/enterprise#40052

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>